Blog-Example created with Bootstrap
This project is a simple blog website built using HTML, CSS, and Bootstrap. It features a responsive design that adapts to different screen sizes, as well as a user-friendly interface that makes it easy to create and manage blog posts.

The blog website is powered by a SQLite database, which stores all of the blog posts, comments, and other user data. The website also uses a Javascript to process user requests and generate the appropriate HTML output.

This project was created as a learning exercise to learn how to build a simple blog website using HTML, CSS, and Bootstrap. It is also a good example of how to use a database to store user data and how to use a Javascript to process user requests.

If you are interested in learning more about how to build a blog website, this project is a good starting point. You can also find additional resources on the Bootstrap website.

The features of your blog website, such as the ability to create and manage blog posts, comment on posts, and subscribe to other users' blogs.
The technologies that you used to build your blog website, such as HTML, CSS, Bootstrap, SQLite, and Javascript.
The challenges that you faced while building your blog website and how you overcame them.
Any additional information that you think would be helpful to other developers who are interested in building their own blog websites.